Queries.ts file line reduced
