+++ src/GraphQl/Queries/Queries.ts
@@ -1,7 +1,20 @@
 import gql from 'graphql-tag';
 
 //Query List
+export { PLUGIN_GET } from './PlugInQueries';
+export { ADVERTISEMENTS_GET } from './PlugInQueries';
+export { ORGANIZATION_EVENTS_CONNECTION } from './PlugInQueries';
+export { USER_TASKS_LIST } from './PlugInQueries';
+export { DIRECT_CHATS_LIST } from './PlugInQueries';
+export { IS_SAMPLE_ORGANIZATION_QUERY } from './PlugInQueries';
+export { ORGANIZATION_CUSTOM_FIELDS } from './PlugInQueries';
 
+export { ORGANIZATION_POST_LIST } from './OrganizationQueries';
+export { ORGANIZATION_POST_CONNECTION_LIST } from './OrganizationQueries';
+export { USER_ORGANIZATION_CONNECTION } from './OrganizationQueries';
+export { USER_JOINED_ORGANIZATIONS } from './OrganizationQueries';
+export { USER_CREATED_ORGANIZATIONS } from './OrganizationQueries';
+export { ORGANIZATION_ADMINS_LIST } from './OrganizationQueries';
 // Check Auth
 
 export const CHECK_AUTH = gql`
@@ -553,311 +566,6 @@ export const MEMBERSHIP_REQUEST = gql`
           email
         }
       }
-    }
-  }
-`;
-
-// display posts
-
-export const ORGANIZATION_POST_LIST = gql`
-  query PostsByOrganization($id: ID!) {
-    postsByOrganization(id: $id) {
-      _id
-      title
-      text
-      imageUrl
-      videoUrl
-      creator {
-        _id
-        firstName
-        lastName
-        email
-      }
-      createdAt
-    }
-  }
-`;
-
-export const ORGANIZATION_POST_CONNECTION_LIST = gql`
-  query PostsByOrganizationConnection(
-    $id: ID!
-    $title_contains: String
-    $text_contains: String
-  ) {
-    postsByOrganizationConnection(
-      id: $id
-      where: { title_contains: $title_contains, text_contains: $text_contains }
-      orderBy: createdAt_DESC
-    ) {
-      edges {
-        _id
-        title
-        text
-        imageUrl
-        videoUrl
-        creator {
-          _id
-          firstName
-          lastName
-          email
-        }
-        createdAt
-        likeCount
-        commentCount
-        comments {
-          _id
-          creator {
-            _id
-            firstName
-            lastName
-            email
-          }
-          likeCount
-          likedBy {
-            _id
-          }
-          text
-        }
-        likedBy {
-          _id
-          firstName
-          lastName
-        }
-        pinned
-      }
-    }
-  }
-`;
-
-export const USER_ORGANIZATION_CONNECTION = gql`
-  query organizationsConnection(
-    $first: Int
-    $skip: Int
-    $filter: String
-    $id: ID
-  ) {
-    organizationsConnection(
-      first: $first
-      skip: $skip
-      where: { name_contains: $filter, id: $id }
-      orderBy: name_ASC
-    ) {
-      _id
-      name
-      image
-      description
-      userRegistrationRequired
-      creator {
-        firstName
-        lastName
-      }
-    }
-  }
-`;
-
-export const USER_JOINED_ORGANIZATIONS = gql`
-  query UserJoinedOrganizations($id: ID!) {
-    users(where: { id: $id }) {
-      joinedOrganizations {
-        _id
-        name
-        description
-        image
-      }
-    }
-  }
-`;
-
-export const USER_CREATED_ORGANIZATIONS = gql`
-  query UserJoinedOrganizations($id: ID!) {
-    users(where: { id: $id }) {
-      createdOrganizations {
-        _id
-        name
-        description
-        image
-      }
-    }
-  }
-`;
-
-export const ORGANIZATION_ADMINS_LIST = gql`
-  query Organizations($id: ID!) {
-    organizations(id: $id) {
-      _id
-      admins {
-        _id
-        image
-        firstName
-        lastName
-        email
-      }
-    }
-  }
-`;
-
-/**
- * {@label PLUGIN_GET}
- * @remarks
- * used to fetch list of plugins
- */
-export const PLUGIN_GET = gql`
-  query getPluginList {
-    getPlugins {
-      _id
-      pluginName
-      pluginCreatedBy
-      pluginDesc
-      uninstalledOrgs
-    }
-  }
-`;
-export const ADVERTISEMENTS_GET = gql`
-  query getAdvertisement {
-    getAdvertisements {
-      _id
-      name
-      type
-      orgId
-      link
-      endDate
-      startDate
-    }
-  }
-`;
-export const ORGANIZATION_EVENTS_CONNECTION = gql`
-  query EventsByOrganizationConnection(
-    $organization_id: ID!
-    $title_contains: String
-    $description_contains: String
-    $location_contains: String
-    $first: Int
-    $skip: Int
-  ) {
-    eventsByOrganizationConnection(
-      where: {
-        organization_id: $organization_id
-        title_contains: $title_contains
-        description_contains: $description_contains
-        location_contains: $location_contains
-      }
-      first: $first
-      skip: $skip
-    ) {
-      _id
-      title
-      description
-      startDate
-      endDate
-      location
-      startTime
-      endTime
-      allDay
-      recurring
-      isPublic
-      isRegisterable
-      creator {
-        _id
-        firstName
-        lastName
-      }
-      attendees {
-        _id
-      }
-    }
-  }
-`;
-
-export const USER_TASKS_LIST = gql`
-  query User($id: ID!) {
-    user(id: $id) {
-      _id
-      assignedTasks {
-        _id
-        title
-        description
-        deadline
-        volunteers {
-          _id
-          firstName
-          lastName
-          email
-        }
-        createdAt
-        completed
-        event {
-          _id
-          title
-          organization {
-            _id
-            name
-            image
-          }
-        }
-        creator {
-          _id
-          firstName
-          lastName
-        }
-      }
-    }
-  }
-`;
-
-export const DIRECT_CHATS_LIST = gql`
-  query DirectChatsByUserID($id: ID!) {
-    directChatsByUserID(id: $id) {
-      _id
-      creator {
-        _id
-        firstName
-        lastName
-        email
-      }
-      messages {
-        _id
-        createdAt
-        messageContent
-        receiver {
-          _id
-          firstName
-          lastName
-          email
-        }
-        sender {
-          _id
-          firstName
-          lastName
-          email
-        }
-      }
-      organization {
-        _id
-        name
-      }
-      users {
-        _id
-        firstName
-        lastName
-        email
-        image
-      }
-    }
-  }
-`;
-
-export const IS_SAMPLE_ORGANIZATION_QUERY = gql`
-  query ($isSampleOrganizationId: ID!) {
-    isSampleOrganization(id: $isSampleOrganizationId)
-  }
-`;
-
-export const ORGANIZATION_CUSTOM_FIELDS = gql`
-  query ($customFieldsByOrganizationId: ID!) {
-    customFieldsByOrganization(id: $customFieldsByOrganizationId) {
-      _id
-      type
-      name
     }
   }
 `;
